+++ unit_tests/util/coordinate.cpp
@@ -40,11 +40,11 @@ BOOST_AUTO_TEST_CASE(regression_test_1347)
     FixedPointCoordinate v(10.001 * COORDINATE_PRECISION, -100.002 * COORDINATE_PRECISION);
     FixedPointCoordinate q(10.002 * COORDINATE_PRECISION, -100.001 * COORDINATE_PRECISION);
 
-    float d1 = coordinate_calculation::perpendicular_distance(u, v, q);
+    double d1 = coordinate_calculation::perpendicular_distance(u, v, q);
 
-    float ratio;
+    double ratio;
     FixedPointCoordinate nearest_location;
-    float d2 = coordinate_calculation::perpendicular_distance(u, v, q, nearest_location, ratio);
+    double d2 = coordinate_calculation::perpendicular_distance(u, v, q, nearest_location, ratio);
 
-    BOOST_CHECK_LE(std::abs(d1 - d2), 0.01f);
+    BOOST_CHECK_LE(std::abs(d1 - d2), 0.01);
 }
